+++ spec/support/examples/test_machine_class.rb
@@ -41,3 +41,10 @@ class ComplexTextMachine
     @fired
   end
 end
+
+class AnotherCustomDefaultMachine
+  include Celluloid::FSM
+
+  state :foo, default: true
+  state :bar
+end